making changes and anew method called show_spinner

+++ lib/tasks/dev.rake
@@ -1,26 +1,31 @@
 namespace :dev do
   desc "Configura o ambiente de desenvolvimento com (:drop :create :migrate :seed)"
   task setup: :environment do
-    spinner = TTY::Spinner.new("[:Spinner] Deletando BD ...")
-    spinner.auto_spin
-    %x(rails db:drop)
-    spinner.success('(concluido!)')
+    if Rails.env.development?
+      show_spinner("Deletando BD ...") do
+        %x(rails db:drop)
+      end
 
-    spinner = TTY::Spinner.new("[:Spinner] Criando BD ...")
-    spinner.auto_spin
-    %x(rails db:create)
-    spinner.success('(concluido!)')
+      show_spinner("Criando BD ...") do
+        %x(rails db:create)
+      end
+      
+      show_spinner("Migrando BD ...") do
+        %x(rails db:migrate)
+      end
 
-    spinner = TTY::Spinner.new("[:Spinner] Migrando BD ... ")
-    spinner.auto_spin
-    %x(rails db:migrate)
-    spinner.success('(concluido!)')
+      show_spinner("Populando BD ...") do
+        %x(rails db:seed)
+      end  
+    else
+      puts "voce nao esta em ambiente rails de desenvolvimento!"
+    end
 
-    spinner = TTY::Spinner.new("[:Spinner] Populando BD ...")
-    spinner.auto_spin
-    %x(rails db:seed)
-    spinner.success('(concluido!)')
- 
+    def show_spinner(msg_start, msg_end = "concluído")
+      spinner = TTY::Spinner.new("[:spinner] #{msg_start} ")
+      spinner.auto_spin
+      %x(rails db:migrate)
+      spinner.success("#{msg_end}")
+    end
   end
-
 end
